Remove debug tree
This feature has served its purpose. It's better to use IPC now.
